% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{add_user_to_map}
\alias{add_user_to_map}
\title{Add user to Maps}
\usage{
add_user_to_map(
  mapfile = "",
  mapuser = "",
  server = Sys.getenv("SUSO_SERVER"),
  workspace = Sys.getenv("SUSO_WORKSPACE"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{mapfile}{Character. Full name of map file as uploaded to workspace and listed in \href{https://docs.mysurvey.solutions/headquarters/preloading/survey-setup-tab-import-copy-and-delete-questionnaire-templates-and-create-assignments/#maps}{HQ-Maps}. Include file extension ('.tpk', '.mmpk' or '.tiff')}

\item{mapuser}{Character. Full name of interviewer account to be added to \code{mapfile}}

\item{server}{Character. Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{workspace}{Character. Name of the workspace in which map is located. In workspace list, value of \code{NAME}, not \verb{DISPLAY NAME}, for the target workspace.}

\item{user}{Character. API or admin user name for user that access to the workspace.}

\item{password}{Character. API or admin password}
}
\value{
Data frame of map and users linked to it
}
\description{
Add an interviewer account ('User') to a map uploaded to workspace
}
\details{
GraphQL implementation of \code{addUserToMap(...)} HeadquartersMutation endpoint.
}
