% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignments.R
\name{archive_assignment}
\alias{archive_assignment}
\title{Archive an assignment}
\usage{
archive_assignment(
  id,
  verbose = FALSE,
  workspace = "primary",
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{id}{Assignment ID number}

\item{verbose}{Returns information about success of operation: message and \code{TRUE}/\code{FALSE} return value.}

\item{workspace}{Character. Name of the workspace whose assignments to get.}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{API user name}

\item{password}{API password}
}
\value{
If \code{verbose = FALSE}, no return value. If \code{verbose = TRUE}, return logical outcome: \code{TRUE} if succeeded; \code{FALSE} otherwise
}
\description{
Wrapper for \code{PATCH /api/v1/assignments/{id}/archive} endpoint
}
