% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{archive_user}
\alias{archive_user}
\title{Archives interviewer or supervisor with all his interviewers}
\usage{
archive_user(
  user_id,
  verbose = FALSE,
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{user_id}{User ID. GUID from server.}

\item{verbose}{Logical. If \code{verbose == TRUE}, return logical outcome and print message. Otherwise, be silent.}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{API user name}

\item{password}{API password}
}
\value{
Side-effect on the server: APPROVE interview. If \code{verbose == TRUE}, return logical outcome and print message.
}
\description{
Wrapper for \code{PATCH /api/v1/users/{id}/archive} endpoint
}
