% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interviews.R
\name{assign_interview_to_sup}
\alias{assign_interview_to_sup}
\title{Reassign an interview to a supervisor}
\usage{
assign_interview_to_sup(
  interview_id,
  user_id = "",
  user_name = "",
  verbose = FALSE,
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{interview_id}{Interview ID. GUID from server or \code{interview__id} from exported data}

\item{user_id}{ID of target user. GUID from server.}

\item{user_name}{User name of target user.}

\item{verbose}{Logical. If \code{verbose = TRUE}, return logical outcome and print message. Otherwise, be silent.}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{API user name}

\item{password}{API password}
}
\value{
If \code{verbose = FALSE}, no return value. If \code{verbose = TRUE}, return logical outcome.
}
\description{
Has the server-side side-effect of reassigning a single interview to a supervisor. For interactive use, prints a message about the outcome. For programmatic use, if \code{verbose = TRUE}, returns a logical value about whether interview successfully assigned, which may be useful for functionst that need to track success and failure of operations.
}
\details{
Wrapper for \code{PATCH ​/api​/v1​/interviews​/{id}​/assignsupervisor} endpoint
}
