% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{assign_users_to_workspaces}
\alias{assign_users_to_workspaces}
\title{Manager user assignment to a workspace}
\usage{
assign_users_to_workspaces(
  users,
  workspaces,
  action = "Assign",
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{users}{Character vector. GUID for user.}

\item{workspaces}{Character vector. Name ID of workspace}

\item{action}{Character. One of the following: Assign, Add, Remove.}

\item{server}{Character. Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{Character. Admin user name}

\item{password}{Character. Admin password}
}
\description{
Manage which users are present in which workspaces. For now, only headquarters and API users may be managed. Actions available: assign, add, and remove. The assign action moves a user to the target workspace (exclusively). The add action adds the user to the workspace while keeping them in previously assigned workspaces. The remove action removes them from the target workspace.
}
\details{
Wrapper for \verb{POST ​/api​/v1​/workspaces​/assign} endpoint.
}
