% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignments.R
\name{change_assignment_quantity}
\alias{change_assignment_quantity}
\title{Change assignment quantity}
\usage{
change_assignment_quantity(
  id,
  quantity,
  workspace = "primary",
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{id}{Assignment ID number}

\item{quantity}{Numeric.}

\item{workspace}{Character. Name of the workspace whose assignments to get.}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{API user name}

\item{password}{API password}
}
\value{
Server-side side-effect of changing assignment quantity.
}
\description{
Change the quantity for the target assignment.
}
\details{
Wrapper for \verb{PATCH ​/api​/v1​/assignments​/\{id\}​/changeQuantity}
}
