% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignments.R
\name{check_assignment_audio}
\alias{check_assignment_audio}
\title{Gets status of audio recording for provided assignment}
\usage{
check_assignment_audio(
  id,
  workspace = "primary",
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{id}{Assignment ID number}

\item{workspace}{Character. Name of the workspace whose assignments to get.}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{API user name}

\item{password}{API password}
}
\value{
Data frame with \verb{id' and }Enabled\code{columns.}Enabled\code{is}TRUE\verb{if audio is enabled;}FALSE\verb{if not enabled;}NA` if the request fails for any reason.
}
\description{
Informs whether audio audit is enabled for the target assignment. Wrapper for \code{GET ​/api​/v1​/assignments​/{id}​/recordAudio} endpoint
}
