% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignments.R
\name{close_assignment}
\alias{close_assignment}
\title{Close assignment}
\usage{
close_assignment(
  id,
  verbose = FALSE,
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{id}{Assignment ID number}

\item{verbose}{Print whether operation was successful or not}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{API user name}

\item{password}{API password}
}
\value{
If \code{verbose = FALSE}, no return value. If \code{verbose = TRUE}, return logical outcome: \code{TRUE} if succeeded; \code{FALSE} otherwise
}
\description{
Closes assignment by setting Size to the amount of collected interviews. Wrapper for \code{POST /api/v1/assignments/{id}/close} endpoint
}
