% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interviews.R
\name{comment_question}
\alias{comment_question}
\title{Leave a comment on a question}
\usage{
comment_question(
  interview_id,
  variable_name,
  roster_vector = "",
  comment,
  verbose = FALSE,
  server = Sys.getenv("SUSO_SERVER"),
  workspace = Sys.getenv("SUSO_WORKSPACE"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{interview_id}{Character. Interview ID. GUID from server or \code{interview__id} from exported data}

\item{variable_name}{Character. Variable name. User name from Designer.}

\item{roster_vector}{Character. Row code(s) of variable. If a single row code, a single character value (e.g., "102"). If multiple row codes, a character containing a comma-separated list (e.g, "1, 2").}

\item{comment}{Character. Comment to post.}

\item{verbose}{Logical. If \code{verbose == TRUE}, return logical outcome and print message. Otherwise, be silent.}

\item{server}{Character. Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{workspace}{Character. Name of the workspace whose interviews to get. In workspace list, value of \code{NAME}, not \verb{DISPLAY NAME}, for the target workspace.}

\item{user}{Character. API user name}

\item{password}{Character. API password}
}
\value{
If \code{verbose = FALSE}, no return value. If \code{verbose = TRUE}, return logical outcome.
}
\description{
Has server-side side-effect of posting a comment to a target question. To target a question, specify it in \code{variable_name}. If the question is in a roster, specify the row code as a comma-separated set of coordinates in \code{row_number}.
}
\details{
Wrapper for \code{POST /api/v1/interviews/{id}/comment-by-variable/{variable}} endpoint
}
