% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{create_user}
\alias{create_user}
\title{Creates new user with specified role.}
\usage{
create_user(
  role,
  supervisor = "",
  user_name,
  user_password = "",
  full_name = "",
  phone_number = "",
  email = "",
  verbose = FALSE,
  server = Sys.getenv("SUSO_SERVER"),
  workspace = Sys.getenv("SUSO_WORKSPACE"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{role}{Character. Role in data collection. Accepted values: \code{"Supervisor"}, \code{"Interviewer"}}

\item{supervisor}{Character. User name of supervisor. Required for \code{role == "Interviewer"}. Omitted otherwise.}

\item{user_name}{Character. User name.}

\item{user_password}{Character. Password for user account.}

\item{full_name}{Character. Full name of user. (Optional)}

\item{phone_number}{Character. Phone number for user. (Optional)}

\item{email}{Character. Email address for user. (Optional)}

\item{verbose}{Logical. If \code{verbose == TRUE}, return logical outcome and print message. Otherwise, be silent.}

\item{server}{Character. Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{workspace}{Character. Name of the workspace whose users to get. In workspace list, value of \code{NAME}, not \verb{DISPLAY NAME}, for the target workspace.}

\item{user}{Character. API user name}

\item{password}{Character. API password}
}
\value{
Side-effect on the server: APPROVE interview. If \code{verbose == TRUE}, return logical outcome and print message.
}
\description{
Creates a new uers with all the specified properties. Wrapper for \code{POST /api/v1/users} endpoint.
}
