% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{create_workspace}
\alias{create_workspace}
\title{Create a workspace}
\usage{
create_workspace(
  name,
  display_name,
  verbose = FALSE,
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{name}{Character. Name identifier of workspace.}

\item{display_name}{Character. Name displayed in GUI to describe workspace.}

\item{verbose}{Boolean. If \code{TRUE}, return a Boolean value about whether operation succeeded.}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{Admin user name}

\item{password}{Admin password}
}
\value{
Server-side side-effect of creating a workspace.
}
\description{
Create a workspace with name ID \code{name} and name displayed in GUI \code{display_name}.
}
\details{
Note: this function requires the credentials of an admin user.

Wrapper for \verb{POST ​/api​/v1​/workspaces} endpoint.
}
