% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{delete_global_notice}
\alias{delete_global_notice}
\title{Delete the global notice}
\usage{
delete_global_notice(
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{User name}

\item{password}{Password}
}
\value{
Server-side side-effect of deleting the global notice text.
}
\description{
Delete the global notice for workspaces in the user's scope. For admin, this is all workspaces. For an API user, this is all workspaces to which the user has been added.
}
\details{
Wrapper for the \verb{DELETE ​/api​/v1​/settings​/globalnotice} endpoint
}
