% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interviews.R
\name{delete_interview}
\alias{delete_interview}
\title{Delete an interview}
\usage{
delete_interview(
  interview_id,
  server = Sys.getenv("SUSO_SERVER"),
  workspace = Sys.getenv("SUSO_WORKSPACE"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{interview_id}{Character. Interview ID. GUID from server or \code{interview__id} from exported data}

\item{server}{Character. Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{workspace}{Character. Name of the workspace whose interviews to get. In workspace list, value of \code{NAME}, not \verb{DISPLAY NAME}, for the target workspace.}

\item{user}{Character. API user name}

\item{password}{Character. API password}
}
\description{
Delete the target interview.
}
\details{
Wrapper for the \verb{DELETE ​/api​/v1​/interviews​/\{id\}} endpoint.
}
