% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{enable_workspace}
\alias{enable_workspace}
\title{Enable a workspace}
\usage{
enable_workspace(
  name,
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{name}{Character. Name identifier of workspace to enable.}

\item{server}{Character. Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{Character. Admin user name}

\item{password}{Character. Admin password}
}
\description{
Enable a previously disabled workspace with name \code{name}.
}
\details{
Wrapper for the \verb{POST ​/api​/v1​/workspaces​/\{name\}​/enable} endpoint.
}
