% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignments.R
\name{get_assignment_details}
\alias{get_assignment_details}
\title{Get details for a single assignment}
\usage{
get_assignment_details(
  id,
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{id}{Integer. Assignment ID}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})server}

\item{user}{API user name}

\item{password}{API password}
}
\value{
Data frame of assignment details for single assignment
}
\description{
Wrapper for \code{GET /api/v1/assignments/{id}} endpoint
}
