% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignments_gql.R
\name{get_assignments_gql}
\alias{get_assignments_gql}
\title{Get all assignments}
\usage{
get_assignments_gql(
  qnr_id = "",
  qnr_version = NULL,
  nodes = c("archived", "createdAtUtc", "email", "id", "interviewsNeeded",
    "receivedByTabletAtUtc", "responsibleId", "webMode", "calendarEvent"),
  assignment.ids = NULL,
  archived = FALSE,
  chunk_size = 100,
  server = Sys.getenv("SUSO_SERVER"),
  workspace = Sys.getenv("SUSO_WORKSPACE"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{qnr_id}{Questionnaire ID. GUID provided by the server.}

\item{qnr_version}{Questionnaire version. Version number provided by the server.}

\item{nodes}{Character vector. Names of attributes to fetch for each map}

\item{assignment.ids}{List with named object 'in' or 'nin' which contains assignment ids to be filtered.}

\item{archived}{Query archived or non-archived assignments}

\item{chunk_size}{Numeric. Number of records to take in one request.}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{workspace}{Character. Name of the workspace whose assignments to get. In workspace list, value of \code{NAME}, not \verb{DISPLAY NAME}, for the target workspace.}

\item{user}{User name}

\item{password}{Password}
}
\value{
Data frame of all assignments that meet search criteria.
}
\description{
Get all assignments for query parameters. (Preliminary) GraphQL implementation of \code{assignments(...)} endpoint.
}
