% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{get_export_file}
\alias{get_export_file}
\title{Downloads export file.}
\usage{
get_export_file(
  job_id,
  path,
  verbose = FALSE,
  server = Sys.getenv("SUSO_SERVER"),
  workspace = Sys.getenv("SUSO_WORKSPACE"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{job_id}{Numeric. Export job ID}

\item{path}{Character. File path where export file should be downloaded}

\item{verbose}{Logical. If \code{verbose = TRUE}, return a logical value about the outcome. If \code{verbose = FALSE}, the default, simply return a message describing the outcome.}

\item{server}{Character. Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{workspace}{Character. Name of the workspace whose export files to get. In workspace list, value of \code{NAME}, not \verb{DISPLAY NAME}, for the target workspace.}

\item{user}{Character. API user name}

\item{password}{Character. API password}
}
\value{
Logical. \code{TRUE} if file successfully downloaded; \code{FALSE} otherwise.
}
\description{
If the export file is ready, this function downloads it to the target directory specified in \code{path}. If the file is not ready, print a message describing the situation.
}
\details{
Wrapper for \code{GET /api/v2/export/{id}/file} endpoint
}
