% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{get_export_job_details}
\alias{get_export_job_details}
\title{Get detailed information about export process}
\usage{
get_export_job_details(
  job_id,
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{job_id}{Export job ID}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{API user name}

\item{password}{API password}
}
\value{
If job found, data frame of export job details. Otherwise, message about why the job was not found.
}
\description{
Get data frame of export job details for a single job. This provides a nice mechanism to inquire about the progress of recently started export jobs and see if they are complete.
}
\details{
Wrapper for \code{GET /api/v2/export/{id}} endpoint
}
