% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{get_export_jobs}
\alias{get_export_jobs}
\title{Get list of export jobs}
\usage{
get_export_jobs(
  export_type = "",
  interview_status = "All",
  qnr_id = "",
  export_status = "",
  has_file = NA,
  limit = 40,
  offset = 0,
  server = Sys.getenv("SUSO_SERVER"),
  workspace = Sys.getenv("SUSO_WORKSPACE"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{export_type}{Character. Type of data to export. Values: \code{c("Tabular", "STATA", "SPSS", "Binary", "DDI", "Binary")}}

\item{interview_status}{Character. Status of interviews to export. Values: \code{c("All", "SupervisorAssigned", "InterviewerAssigned", "Completed", "RejectedBySupervisor", "ApprovedBySupervisor", "RejectedByHeadquarters", "ApprovedByHeadquarters")}}

\item{qnr_id}{Character. Questionnaire ID. Format: \code{qnr_id$version}}

\item{export_status}{Character. Status of export process. Values: \code{c("Created", "Running", "Completed", "Fail", "Canceled")}}

\item{has_file}{Logical. Whether export file generated.}

\item{limit}{Number of results to return. Values: \code{c(1:40)}}

\item{offset}{Offset in list of processes.}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{workspace}{Character. Name of the workspace whose export jobs to get. In workspace list, value of \code{NAME}, not \verb{DISPLAY NAME}, for the target workspace.}

\item{user}{API user name}

\item{password}{API password}
}
\value{
Data frame of export jobs.
}
\description{
Get data frame of export jobs on the server. This includes jobs that have been started, cancelled, and completed. The data frame includes the job id as well as several attributes about the job.
}
\details{
Wrapper for \code{GET /api/v2/export} endpoint
}
