% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{get_global_notice}
\alias{get_global_notice}
\title{Get the global notice}
\usage{
get_global_notice(
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{User name}

\item{password}{Password}
}
\value{
Character. Global notice.
}
\description{
Return the global notice, if any, set for workspaces within the user's scope. For admin, this is all workspaces. For an API user, this is all workspaces to which the user has been added.
}
\details{
Wrapper for the \verb{GET ​/api​/v1​/settings​/globalnotice} endpoint.
}
