% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interviews.R
\name{get_interview_history}
\alias{get_interview_history}
\title{Get interview history}
\usage{
get_interview_history(
  interview_id,
  server = Sys.getenv("SUSO_SERVER"),
  workspace = Sys.getenv("SUSO_WORKSPACE"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{interview_id}{Character. GUID for interview.}

\item{server}{Character. Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{workspace}{Character. Name of the workspace whose interviews to get. In workspace list, value of \code{NAME}, not \verb{DISPLAY NAME}, for the target workspace.}

\item{user}{Character. API user name}

\item{password}{Character. API password}
}
\value{
List
}
\description{
Get the history of interview actions for the target interview.
}
\details{
Wrapper for the \verb{GET ​/api​/v1​/interviews​/\{id\}​/history} endpoint.
}
