% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interviews.R
\name{get_interview_stats}
\alias{get_interview_stats}
\title{Get statistics by interview}
\usage{
get_interview_stats(
  interview_id,
  workspace = "primary",
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{interview_id}{Interview ID. GUID from server or \code{interview__id} from exported data}

\item{workspace}{Character. Name of the workspace whose interviews to get.}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{API user name}

\item{password}{API password}
}
\value{
Data frame with all values provided by the server.
}
\description{
Fetch statistics for a single interview.
}
\details{
These statistics provide metadata about the interview process. Columns include: \code{Answered}, \code{NotAnswered}, \code{Flagged}, \code{NotFlagged}, \code{Valid}, \code{Invalid}, \code{WithComments}, \code{ForInterviewer}, \code{ForSupervisor}, \code{InterviewId}, \code{InterviewKey}, \code{Status}, \code{ResponsibleId}, \code{ResponsibleName}, \code{NumberOfInterviewers}, \code{NumberRejectionsBySupervisor}, \code{NumberRejectionsByHq}, \code{InterviewDuration}, \code{AssignmentId}, \code{UpdatedAtUtc}.

Wrapper for \code{GET /api/v1/interviews/{id}/stats} endpoint
}
