% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_interviewers}
\alias{get_interviewers}
\title{Get list of all interviewers}
\usage{
get_interviewers(
  workspace = "primary",
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{workspace}{Character. Name of the workspace whose users to get.}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{API user name}

\item{password}{API password}
}
\value{
Data frame of interviewers and their details.
}
\description{
Fetch list of all interviewers and their attributes, including their supervisor
}
