% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interviews.R
\name{get_interviews}
\alias{get_interviews}
\title{Get interviews}
\usage{
get_interviews(
  nodes = c("id", "key", "assignmentId", "identifyingData", "questionnaireId",
    "questionnaireVersion", "questionnaireVariable", "responsibleName", "responsibleId",
    "responsibleRole", "supervisorName", "status", "actionFlags", "wasCompleted",
    "notAnsweredCount", "errorsCount", "createdDate", "updateDateUtc",
    "receivedByInterviewerAtUtc", "interviewMode"),
  chunk_size = 100,
  server = Sys.getenv("SUSO_SERVER"),
  workspace = Sys.getenv("SUSO_WORKSPACE"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{nodes}{Character vector. Names of attributes to fetch for each interview.}

\item{chunk_size}{Numeric. Number of records to take in one request.}

\item{server}{Character. Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{workspace}{Character. Name of the workspace whose interviews to get. In workspace list, value of \code{NAME}, not \verb{DISPLAY NAME}, for the target workspace.}

\item{user}{Character. API or admin user name for user that access to the workspace.}

\item{password}{Character. API or admin password}
}
\value{
Data frame of interviews and their (user-specified) attributes.
}
\description{
Get list of interviews and their attributes
}
\details{
GraphQL implementation of deprecated REST \verb{GET ​/api​/v1​/interviews​/\{id\}} endpoint.
}
