% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{get_maps}
\alias{get_maps}
\title{Get maps}
\usage{
get_maps(
  nodes = c("fileName", "size", "users", "importDateUtc", "xMaxVal", "yMaxVal",
    "xMinVal", "yMinVal", "wkid", "maxScale", "minScale"),
  chunk_size = 100,
  server = Sys.getenv("SUSO_SERVER"),
  workspace = Sys.getenv("SUSO_WORKSPACE"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{nodes}{Character vector. Names of attributes to fetch for each map}

\item{chunk_size}{Numeric. Number of records to take in one request.}

\item{server}{Character. Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{workspace}{Character. Name of the workspace whose maps to get. In workspace list, value of \code{NAME}, not \verb{DISPLAY NAME}, for the target workspace.}

\item{user}{Character. API or admin user name for user that access to the workspace.}

\item{password}{Character. API or admin password}
}
\value{
Data frame of maps and their (user-specified) attributes.
}
\description{
Get list of maps uploaded to workspace and their attributes
}
\details{
GraphQL implementation of \code{maps(...)} endpoint.
}
