% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaires.R
\name{get_possible_interview_statuses}
\alias{get_possible_interview_statuses}
\title{Get possible interview statuses}
\usage{
get_possible_interview_statuses(
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{server}{Character. Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{Character. API user name}

\item{password}{Character. API password}
}
\value{
Character vector. Names of all possible interview statuses
}
\description{
Wrapper for the \verb{GET ​/api​/v1​/questionnaires​/statuses} endpoint.
}
