% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaires.R
\name{get_questionnaire_document}
\alias{get_questionnaire_document}
\title{Save JSON representation of the questionnaire}
\usage{
get_questionnaire_document(
  qnr_id,
  qnr_version,
  path,
  server = Sys.getenv("SUSO_SERVER"),
  workspace = Sys.getenv("SUSO_WORKSPACE"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{qnr_id}{Character. Questionnaire ID. GUID from server}

\item{qnr_version}{Numeric. Version number of questionnaire}

\item{path}{Character. Directory where export JSON representation of the questionnaire should be downloaded}

\item{server}{Character. Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{workspace}{Character. Name of the workspace whose questionnaire document to get. In workspace list, value of \code{NAME}, not \verb{DISPLAY NAME}, for the target workspace.}

\item{user}{Character. API user name}

\item{password}{Character. API password}
}
\description{
Fetches JSON representation of the questionnaire. Save it to disk. Wrapper for \code{GET ​/api​/v1​/questionnaires​/{id}​/{version}​/document} endpoint
}
