% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_server_health}
\alias{get_server_health}
\title{Get the server health status}
\usage{
get_server_health(server = Sys.getenv("SUSO_SERVER"))
}
\arguments{
\item{server}{Character. Full server web address
(e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}
}
\value{
Character. Survey Solutions version
}
\description{
Extracts server health status from the \code{.hc} sub-page
}
