% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_server_version}
\alias{get_server_version}
\title{Get the Survey Solutions version on the server}
\usage{
get_server_version(server = Sys.getenv("SUSO_SERVER"))
}
\arguments{
\item{server}{Character. Full server web address
(e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}
}
\value{
Character. Survey Solutions version
}
\description{
Extracts software version number from the \code{.version} sub-page
}
