% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_user_action_log}
\alias{get_user_action_log}
\title{Get action log for single user}
\usage{
get_user_action_log(
  user_id,
  start = "",
  end = "",
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{user_id}{User ID. GUID from server.}

\item{start}{Character. Date in "YYYY-MM-DD" format}

\item{end}{Character. Date in "YYYY-MM-DD" format}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{API user name}

\item{password}{API password}
}
\value{
Data frame containing the action log.
}
\description{
Fetches the log of user actions on the tablet. This includes creating, opening, editing, closing, and sending interviews. Returns the log as a data frame.
}
\details{
Wrapper for \code{GET /api/v1/interviewers/{id}/actions-log} endpoint
}
