% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{get_workspace_details}
\alias{get_workspace_details}
\title{Get details of a single workspace}
\usage{
get_workspace_details(
  name,
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{name}{Character. Name identifier of workspace.}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{Admin user name}

\item{password}{Admin password}
}
\value{
Data frame containing workspace details: \code{Name}, the name ID; \code{DisplayName}, the name displayed in the GUI; and \code{DisabledAtUtc}, the time the workspace was disabled or \code{NA} if not disabled
}
\description{
Rather than get the details for all workspaces, get them for a single workspace.
}
\details{
Wrapper for \verb{GET ​/api​/v1​/workspaces​/\{name\}} endpoint
}
