% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{get_workspace_status}
\alias{get_workspace_status}
\title{Get detailed workspace status}
\usage{
get_workspace_status(
  name,
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{name}{Character. Name identifier of workspace to delete.}

\item{server}{Character. Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{Character. Admin or API user name}

\item{password}{Character. Admin or API password}
}
\value{
Data frame. Contains columns: \code{CanBeDeleted}, \code{WorkspaceName}, \code{WorkspaceDisplayName}, \code{ExistingQuestionnairesCount}, \code{InterviewersCount}, \code{SupervisorsCount}, \code{MapsCount}
}
\description{
Obtains a data frame with details about the workspace status and contents
}
\details{
Wrapper for \verb{GET ​/api​/v1​/workspaces​/status​/\{name\}} endpoint.
}
