% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{get_workspaces}
\alias{get_workspaces}
\title{Get list of workspaces}
\usage{
get_workspaces(
  user_id = "",
  include_disabled = FALSE,
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{user_id}{GUID. Searches for workspaces where user with user ID is assigned.}

\item{include_disabled}{Boolean. Determine whether list of workspaces should include those that have been disabled or not.}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{Admin or API user name}

\item{password}{Admin or API user password}
}
\value{
Data frame of workspaces. Contains columns: \code{Name}, the name ID; \code{DisplayName}, the name in the GUI; and \code{DisabledAtUtc}, when the workspace was disabled or NA if not disabled.
}
\description{
Get list of workspaces that meet query criteria. The list is affected by what the user can "see". If the credentials are admin credentials, all workspaces will be returned. If the user is an API user, then only those workspaces where the API user is present, will be included.
}
\details{
Wrapper for the \verb{GET ​/api​/v1​/workspaces} endpoint.
}
