% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_user_name}
\alias{is_user_name}
\title{Checks whether is a valid user name}
\usage{
is_user_name(name)
}
\arguments{
\item{name}{Character. User name to validate.}
}
\description{
Returns a Boolean value
}
