% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignments.R
\name{reassign_assignment}
\alias{reassign_assignment}
\title{Reassign assignment to another user}
\usage{
reassign_assignment(
  id,
  responsible,
  server = Sys.getenv("SUSO_SERVER"),
  workspace = Sys.getenv("SUSO_WORKSPACE"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{id}{Numeric. Assignment ID number}

\item{responsible}{Character. Either user name or GUID.}

\item{server}{Character. Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{workspace}{Character. Name of the workspace whose assignments to get. In workspace list, value of \code{NAME}, not \verb{DISPLAY NAME}, for the target workspace.}

\item{user}{Character. API user name}

\item{password}{Character. API password}
}
\value{
Server-side side-effect of reassigning an assignment to another user
}
\description{
Wrapper for \verb{PATCH ​/api​/v1​/assignments​/\{id\}​/assign}
}
