% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interviews.R
\name{reject_interview_as_sup}
\alias{reject_interview_as_sup}
\title{Reject an interview as supervisor}
\usage{
reject_interview_as_sup(
  interview_id,
  comment = "",
  responsible_id = "",
  verbose = FALSE,
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{interview_id}{Interview ID. GUID from server or \code{interview__id} from exported data}

\item{comment}{Comment to post upon rejection}

\item{responsible_id}{User to receive rejected interview. User ID. GUID from server.}

\item{verbose}{Logical. If \code{verbose == TRUE}, return logical outcome and print message. Otherwise, be silent.}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{API user name}

\item{password}{API password}
}
\value{
If \code{verbose = FALSE}, no return value. If \code{verbose = TRUE}, return logical outcome.
}
\description{
Has server-side side-effect of rejecting a single interview as a supervisor. For interactive use, prints a message to describe the outcome of the action. If \code{verbose = TRUE}, returns a logical value about the outcome that may be useful for functions that need to track successes and failures.
}
\details{
Wrapper for \code{PATCH /api/v1/interviews/{id}/reject} endpoint
}
