% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignments.R
\name{set_assignment_audio}
\alias{set_assignment_audio}
\title{Set whether assignment will record audio or not}
\usage{
set_assignment_audio(
  id,
  enable,
  verbose,
  server = Sys.getenv("SUSO_SERVER"),
  workspace = Sys.getenv("SUSO_WORKSPACE"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{id}{Number. Assignment ID number}

\item{enable}{Logical. Whether to enable or disable audio--\code{TRUE} or \code{FALSE}, respectively}

\item{verbose}{Logical. Returns information about success of operation: message and \code{TRUE}/\code{FALSE} return value.}

\item{server}{Character. Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{workspace}{Character. Name of the workspace whose assignments to get. In workspace list, value of \code{NAME}, not \verb{DISPLAY NAME}, for the target workspace.}

\item{user}{Character. API user name}

\item{password}{Character. API password}
}
\value{
If \code{verbose = FALSE}, no return value. If \code{verbose = TRUE}, return logical outcome: \code{TRUE} if succeeded; \code{FALSE} otherwise
}
\description{
Wrapper for \code{PATCH ​/api​/v1​/assignments​/{id}​/recordAudio} endpoint
}
