% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{set_credentials}
\alias{set_credentials}
\title{Set server credentials}
\usage{
set_credentials(server, user, password)
}
\arguments{
\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{API user name}

\item{password}{API password}
}
\value{
Environment variables \code{SUSO_SERVER}, \code{SUSO_USER}, \code{SUSO_PASSWORD}
}
\description{
Stores server access credentials for later use. For the current session, credentials are accessible in environment variables. For future sessions, credentials are stored in the .Renviron and loaded upon startup. While functions in this package automatically load the credential set in this way, the credentials can also be accessed via the \code{Sys.getenv} function and using the credential name SUSO_SERVER (server address), SUSO_USER (API user name), SUSO_PASSWORD (API user password)
}
