% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{set_global_notice}
\alias{set_global_notice}
\title{Set the global notice}
\usage{
set_global_notice(
  text,
  server = Sys.getenv("SUSO_SERVER"),
  workspace = Sys.getenv("SUSO_WORKSPACE"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{text}{Character. Text of the global notice to display.}

\item{server}{Character. Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{workspace}{Character. Name of the workspace whose global notice to get. In workspace list, value of \code{NAME}, not \verb{DISPLAY NAME}, for the target workspace.}

\item{user}{Character. User name}

\item{password}{Character. Password}
}
\value{
Server-side side-effect of setting the global notice text.
}
\description{
Set the global notice for workspaces within the user's scope. For admin, this is all workspaces. For an API user, this is all workspaces to which the user has been added.
}
\details{
Wrapper for the \verb{PUT /api​/v1​/settings​/globalnotice} endpoint.
}
