% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaires.R
\name{set_questionnaire_audio}
\alias{set_questionnaire_audio}
\title{Enable audio recording for questionnaire}
\usage{
set_questionnaire_audio(
  qnr_id,
  qnr_version,
  enable,
  server = Sys.getenv("SUSO_SERVER"),
  workspace = Sys.getenv("SUSO_WORKSPACE"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{qnr_id}{Character. Questionnaire ID. GUID from server.}

\item{qnr_version}{Numeric. Questionnaire version number.}

\item{enable}{Logical. Whether to enable. Values: c(TRUE, FALSE)}

\item{server}{Character. Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{workspace}{Character. Name of the workspace whose questionnaire audio settings to change. In workspace list, value of \code{NAME}, not \verb{DISPLAY NAME}, for the target workspace.}

\item{user}{Character. API user name}

\item{password}{Character. API password}
}
\value{
TRUE/FALSE depending on whether operation succeeded or not.
}
\description{
Sets audio recording enabled setting for provided questionnaire. Wrapper of \code{POST /api/v1/questionnaires/{id}/{version}/recordAudio} endpoint
}
