% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{show_credentials}
\alias{show_credentials}
\title{Show saved server credentials}
\usage{
show_credentials()
}
\value{
Side-effect of message in console
}
\description{
Shows server credentials saved in environment variables and .Renviron.
}
