% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{start_export}
\alias{start_export}
\title{Start export file creation}
\usage{
start_export(
  qnr_id,
  export_type,
  interview_status = "All",
  from = "",
  to = "",
  access_token = "",
  refresh_token = "",
  storage_type = "",
  translation_id = "",
  include_meta = TRUE,
  workspace = "primary",
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{qnr_id}{Questionnaire ID. Format: \code{qnr_id$version}}

\item{export_type}{Type of data to export. Values: \code{c("Tabular", "STATA", "SPSS", "Binary", "DDI", "Paradata")}}

\item{interview_status}{Status of interviews to export. Values: \code{c("All", "SupervisorAssigned", "InterviewerAssigned", "Completed", "RejectedBySupervisor", "ApprovedBySupervisor", "RejectedByHeadquarters", "ApprovedByHeadquarters")}}

\item{from}{Start date for interviews to export. date-time, UTC}

\item{to}{End date for interviews to export. date-time, UTC}

\item{access_token}{Access token to external storage. Relevant only if external cloud storage is destination for the export file.}

\item{refresh_token}{Refresh token to external storage.Relevant only if external cloud storage is destination for the export file.}

\item{storage_type}{External storage type, if relevant. Values: \code{c("Dropbox", "OneDrive", "GoogleDrive")}}

\item{translation_id}{Translation ID for variable and value labels to include in export files.}

\item{include_meta}{Logical. If \code{TRUE}, include questionnaire metadata in export file.}

\item{workspace}{Character. Name of the workspace whose export jobs to start.}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{API user name}

\item{password}{API password}
}
\value{
Job ID of export process started.
}
\description{
Has server-side side-effect of starting an export job for the target questionnaire. The function parameters provide a way to describe which interviews should be exported and for which questionnaire.
}
\details{
Wrapper for \code{POST /api/v2/export} endpoint
}
