% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interviews.R
\name{unapprove_interview}
\alias{unapprove_interview}
\title{Unapprove an interview}
\usage{
unapprove_interview(
  interview_id,
  comment = "",
  verbose = FALSE,
  server = Sys.getenv("SUSO_SERVER"),
  workspace = Sys.getenv("SUSO_WORKSPACE"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{interview_id}{Character. Interview ID. GUID from server or \code{interview__id} from exported data}

\item{comment}{Character. Comment to post upon unapproval}

\item{verbose}{Logical. If \code{verbose == TRUE}, return logical outcome and print message. Otherwise, be silent.}

\item{server}{Character. Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{workspace}{Character. Name of the workspace whose interviews to get. In workspace list, value of \code{NAME}, not \verb{DISPLAY NAME}, for the target workspace.}

\item{user}{Character. API user name}

\item{password}{Character. API password}
}
\value{
If \code{verbose = FALSE}, no return value. If \code{verbose = TRUE}, return logical outcome.
}
\description{
Has server-side side-effect of unapproving a single interview, returningn it to Approved by Headquarters status. For interactive use, prints a message to describe the outcome of the action. If \code{verbose = TRUE}, returns a logical value about the outcome that may be useful for functions that need to track successes and failures.
}
\details{
Wrapper for \code{PATCH /api/v1/interviews/{id}/hqunapprove} endpoint
}
