% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{unarchive_user}
\alias{unarchive_user}
\title{Unarchives single user}
\usage{
unarchive_user(
  user_id,
  verbose = FALSE,
  server = Sys.getenv("SUSO_SERVER"),
  workspace = Sys.getenv("SUSO_WORKSPACE"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{user_id}{Character. User ID. GUID from server.}

\item{verbose}{Logical. If \code{verbose == TRUE}, return logical outcome and print message. Otherwise, be silent.}

\item{server}{Character. Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{workspace}{Character. Name of the workspace whose users to get. In workspace list, value of \code{NAME}, not \verb{DISPLAY NAME}, for the target workspace.}

\item{user}{Character. API user name}

\item{password}{Character. API password}
}
\value{
Side-effect on the server: APPROVE interview. If \code{verbose == TRUE}, return logical outcome and print message.
}
\description{
Wrapper for \code{PATCH /api/v1/users/{id}/unarchive} endpoint
}
