% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{update_workspace}
\alias{update_workspace}
\title{Update workspace}
\usage{
update_workspace(
  name,
  display_name,
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{name}{Character. Name identifier of workspace to update.}

\item{display_name}{Character. Name displayed in GUI to describe workspace. This value is updated.}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{Admin or API user name}

\item{password}{Admin or API password}
}
\value{
Server-side side-effect of updating workspace
}
\description{
Updates workspace attributes. For the moment, this function/endpoint only updates the display name.
}
\details{
Wrapper for \verb{PATCH ​/api​/v1​/workspaces​/\{name\}} endpoint
}
