% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interviews.R
\name{assign_interview_to_int}
\alias{assign_interview_to_int}
\title{Reassign interview to interviewer}
\usage{
assign_interview_to_int(
  interview_id,
  user_id = "",
  user_name = "",
  verbose = FALSE,
  server = Sys.getenv("SUSO_SERVER"),
  workspace = Sys.getenv("SUSO_WORKSPACE"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{interview_id}{Character. Interview ID. GUID from server or \code{interview__id} from exported data}

\item{user_id}{Character. User ID. GUID from server.}

\item{user_name}{Character. User name of target interviewer.}

\item{verbose}{Logical. If \code{verbose == TRUE}, return logical outcome and print message. Otherwise, be silent.}

\item{server}{Character. Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{workspace}{Character. Name of the workspace whose interviews to get. In workspace list, value of \code{NAME}, not \verb{DISPLAY NAME}, for the target workspace.}

\item{user}{Character. API user name}

\item{password}{Character. API password}
}
\value{
If \code{verbose = FALSE}, no return value. If \code{verbose = TRUE}, return logical outcome.
}
\description{
Has the server-side side-effect of reassigning a single interview to a target interviewer. For interactive use, prints a message to describe the outcome of the action. If \code{verbose = TRUE}, returns a logical value about the outcome that may be useful for functions that need to track successes and failures.
}
\details{
Wrapper for \code{PATCH /api/v1/interviews/{id}/assign} endpoint
}
