% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{cancel_export}
\alias{cancel_export}
\title{Cancel export process}
\usage{
cancel_export(
  job_id,
  verbose = FALSE,
  workspace = "primary",
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{job_id}{Export job ID}

\item{verbose}{Logical. If \code{verbose = TRUE}, return \code{TRUE} if cancellation successful, \code{FALSE} otherwise}

\item{workspace}{Character. Name of the workspace whose export jobs to cancel.}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{API user name}

\item{password}{API password}
}
\value{
If \code{verbose = FALSE}, no return value. If \code{verbose = TRUE}, return logical outcome.
}
\description{
Has server-side side-effect of cancelling an active export process. A message will inform about the outcome. If \code{verbose = TRUE}, the function also returns a logical value: \code{TRUE} if cancellation successful; \code{FALSE} otherwise.
}
\details{
Wrapper for \code{DELETE /api/v2/export/{id}} endpoint
}
