% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{check_credentials}
\alias{check_credentials}
\title{Check that server credentials are valid}
\usage{
check_credentials(verbose = FALSE)
}
\arguments{
\item{verbose}{Logical. If \code{TRUE}, function returns logical regarding validity of credentials. If \code{FALSE}, function simply prints information about validity to the console.}
}
\value{
If \code{verbose = FALSE} (default), side-effect of message printed to the console. If \code{verbose = TRUE}, logical: \code{TRUE} if credentials valid; \code{FALSE} otherwise.
}
\description{
Shows server credentials saved in environment variables and .Renviron.
}
