% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interviews.R
\name{comment_question}
\alias{comment_question}
\title{Leave a comment on a question}
\usage{
comment_question(
  interview_id,
  variable_name,
  row_number = "",
  comment,
  verbose = FALSE,
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{interview_id}{Interview ID. GUID from server or \code{interview__id} from exported data}

\item{variable_name}{Variable name. User name from Designer.}

\item{row_number}{Row number. Use roster-specific row code.}

\item{comment}{Comment to post.}

\item{verbose}{Logical. If \code{verbose == TRUE}, return logical outcome and print message. Otherwise, be silent.}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{API user name}

\item{password}{API password}
}
\value{
If \code{verbose = FALSE}, no return value. If \code{verbose = TRUE}, return logical outcome.
}
\description{
Has server-side side-effect of posting a comment to a target question. To target a question, specify it in \code{variable_name}. If the question is in a roster, specify the row code as a comma-separated set of coordinates in \code{row_number}.
}
\details{
Wrapper for \code{POST /api/v1/interviews/{id}/comment-by-variable/{variable}} endpoint
}
