% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interviews.R
\name{delete_interview}
\alias{delete_interview}
\title{Delete an interview}
\usage{
delete_interview(
  interview_id,
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{interview_id}{Interview ID. GUID from server or \code{interview__id} from exported data}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{API user name}

\item{password}{API password}
}
\description{
Delete the target interview.
}
\details{
Wrapper for the \verb{DELETE ​/api​/v1​/interviews​/\{id\}} endpoint.
}
