% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignments.R
\name{get_assignment_quantity_setting}
\alias{get_assignment_quantity_setting}
\title{Returns whether assignment quantity can be changed or not}
\usage{
get_assignment_quantity_setting(
  id,
  server = Sys.getenv("SUSO_SERVER"),
  workspace = Sys.getenv("SUSO_WORKSPACE"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{id}{Assignment ID}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})server}

\item{workspace}{Character. Name of the workspace whose assginment quantity to get. In workspace list, value of \code{NAME}, not \verb{DISPLAY NAME}, for the target workspace.}

\item{user}{API user name}

\item{password}{API password}
}
\value{
Data frame with \code{id} and \code{CanChangeQuantity} columns
}
\description{
Wrapper for \code{GET ​/api​/v1​/assignments​/{id}​/assignmentQuantitySettings} endpoint
}
