% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{get_export_file}
\alias{get_export_file}
\title{Downloads export file.}
\usage{
get_export_file(
  job_id,
  path,
  verbose = FALSE,
  workspace = "primary",
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{job_id}{Export job ID}

\item{path}{File path where export file should be downloaded}

\item{verbose}{Logical. If \code{verbose = TRUE}, return a logical value about the outcome. If \code{verbose = FALSE}, the default, simply return a message describing the outcome.}

\item{workspace}{Character. Name of the workspace whose export files to get.}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{API user name}

\item{password}{API password}
}
\value{
Logical. \code{TRUE} if file successfully downloaded; \code{FALSE} otherwise.
}
\description{
If the export file is ready, this function downloads it to the target directory specified in \code{path}. If the file is not ready, print a message describing the situation.
}
\details{
Wrapper for \code{GET /api/v2/export/{id}/file} endpoint
}
