% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaires.R
\name{get_interviews_for_questionnaire}
\alias{get_interviews_for_questionnaire}
\title{Get list of interviews for questionnaire-version}
\usage{
get_interviews_for_questionnaire(
  workspace = "primary",
  chunk_size = 100,
  qnr_id,
  qnr_version,
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{workspace}{Character. Name of the workspace whose questionnaires and associated interviews to get.}

\item{chunk_size}{Numeric. Number of records to take in one request.}

\item{qnr_id}{Questionnaire ID. GUID from server.}

\item{qnr_version}{Questionnaire version number.}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{API user name}

\item{password}{API password}
}
\value{
Data frame of interviews.
}
\description{
GraphQL implmentation for deprecated REST \code{GET /api/v1/questionnaires/{id}/{version}/interviews} endpoint
}
