% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaires.R
\name{get_questionnaire_document}
\alias{get_questionnaire_document}
\title{Save JSON representation of the questionnaire}
\usage{
get_questionnaire_document(
  qnr_id,
  qnr_version,
  path,
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{qnr_id}{Questionnaire ID. GUID from server}

\item{qnr_version}{Version number of questionnaire}

\item{path}{Directory where export JSON representation of the questionnaire should be downloaded}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{API user name}

\item{password}{API password}
}
\description{
Fetches JSON representation of the questionnaire. Save it to disk. Wrapper for \code{GET ​/api​/v1​/questionnaires​/{id}​/{version}​/document} endpoint
}
