% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaires.R
\name{get_questionnaires}
\alias{get_questionnaires}
\title{Get all questionnaires}
\usage{
get_questionnaires(
  server = Sys.getenv("SUSO_SERVER"),
  workspace = Sys.getenv("SUSO_WORKSPACE"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{server}{Character. Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{workspace}{Character. Name of the workspace whose questionnaires to get. In workspace list, value of \code{NAME}, not \verb{DISPLAY NAME}, for the target workspace.}

\item{user}{Character. API user name}

\item{password}{Character. API password}
}
\value{
Data frame of questionnaires.
}
\description{
Get list of all questionnaires and their attributes
}
\details{
GraphQL implementation of the deprecated REST \verb{GET​/api​/v1​/questionnaires} endpoint.
}
