% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaires.R
\name{get_questionnaires}
\alias{get_questionnaires}
\title{Get all questionnaires}
\usage{
get_questionnaires(
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{API user name}

\item{password}{API password}
}
\value{
Data frame of questionnaires.
}
\description{
Get list of all questionnaires and their attributes
}
