% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_supervisors}
\alias{get_supervisors}
\title{Get list of all supervisors}
\usage{
get_supervisors(
  server = Sys.getenv("SUSO_SERVER"),
  workspace = Sys.getenv("SUSO_WORKSPACE"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{server}{Character. Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{workspace}{Character. Name of the workspace whose users to get. In workspace list, value of \code{NAME}, not \verb{DISPLAY NAME}, for the target workspace.}

\item{user}{Character. API user name}

\item{password}{Character. API password}
}
\value{
Data frame of supervisors and their attributes.
}
\description{
Fetch list of all supervisors and their attributes
}
