% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_user_details}
\alias{get_user_details}
\title{Gets detailed info about single user}
\usage{
get_user_details(
  user_id,
  server = Sys.getenv("SUSO_SERVER"),
  workspace = Sys.getenv("SUSO_WORKSPACE"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{user_id}{Character. User ID, user name, or email.}

\item{server}{Character. Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{workspace}{Character. Name of the workspace whose users to get. In workspace list, value of \code{NAME}, not \verb{DISPLAY NAME}, for the target workspace.}

\item{user}{Character. API user name}

\item{password}{Character. API password}
}
\value{
Data frame of single user's details.
}
\description{
Wrapper for \code{GET /api/v1/users/{id}} endpoint
}
