% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignments.R
\name{reassign_assignment}
\alias{reassign_assignment}
\title{Reassign assignment to another user}
\usage{
reassign_assignment(
  id,
  responsible,
  workspace = "primary",
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{id}{Assignment ID number}

\item{responsible}{Character. Either user name or GUID.}

\item{workspace}{Character. Name of the workspace whose assignments to get.}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{API user name}

\item{password}{API password}
}
\value{
Server-side side-effect of reassigning an assignment to another user
}
\description{
Wrapper for \verb{PATCH ​/api​/v1​/assignments​/\{id\}​/assign}
}
