% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interviews.R
\name{reject_interview_as_hq}
\alias{reject_interview_as_hq}
\title{Reject an interview as headquarters}
\usage{
reject_interview_as_hq(
  interview_id,
  comment = "",
  responsible_id = "",
  verbose = FALSE,
  server = Sys.getenv("SUSO_SERVER"),
  workspace = Sys.getenv("SUSO_WORKSPACE"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{interview_id}{Character. Interview ID. GUID from server or \code{interview__id} from exported data.}

\item{comment}{Character. Comment to post upon rejection.}

\item{responsible_id}{Character. User ID. GUID from server or \code{interview__id} from exported data. User ID of the interviewer/supervisor to which the interview should be rejected.}

\item{verbose}{Logical. If \code{verbose == TRUE}, return logical outcome and print message. Otherwise, be silent.}

\item{server}{Character. Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{workspace}{Character. Name of the workspace whose interviews to get. In workspace list, value of \code{NAME}, not \verb{DISPLAY NAME}, for the target workspace.}

\item{user}{Character. API user name}

\item{password}{Character. API password}
}
\value{
If \code{verbose = FALSE}, no return value. If \code{verbose = TRUE}, return logical outcome.
}
\description{
Has server-side side-effect of rejecting a single interview as Headquarters. If \code{verbose = TRUE}, return logical outcome and print message.
}
\details{
Wrapper for \code{PATCH /api/v1/interviews/{id}/hqreject} endpoint
}
