% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignments.R
\name{set_assignment_audio}
\alias{set_assignment_audio}
\title{Set whether assignment will record audio or not}
\usage{
set_assignment_audio(
  id,
  enable,
  verbose,
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{id}{Assignment ID number}

\item{enable}{Whether to enable or disable audio--\code{TRUE} or \code{FALSE}, respectively}

\item{verbose}{Returns information about success of operation: message and \code{TRUE}/\code{FALSE} return value.}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{API user name}

\item{password}{API password}
}
\value{
If \code{verbose = FALSE}, no return value. If \code{verbose = TRUE}, return logical outcome: \code{TRUE} if succeeded; \code{FALSE} otherwise
}
\description{
Wrapper for \code{PATCH ​/api​/v1​/assignments​/{id}​/recordAudio} endpoint
}
