% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{set_global_notice}
\alias{set_global_notice}
\title{Set the global notice}
\usage{
set_global_notice(
  text,
  workspace = "primary",
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{text}{Character. Text of the global notice to display.}

\item{workspace}{Character. Name of the workspace whose global notice to get.}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{User name}

\item{password}{Password}
}
\value{
Server-side side-effect of setting the global notice text.
}
\description{
Set the global notice for workspaces within the user's scope. For admin, this is all workspaces. For an API user, this is all workspaces to which the user has been added.
}
\details{
Wrapper for the \verb{PUT /api​/v1​/settings​/globalnotice} endpoint.
}
