% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaires.R
\name{set_questionnaire_audio}
\alias{set_questionnaire_audio}
\title{Enable audio recording for questionnaire}
\usage{
set_questionnaire_audio(
  qnr_id,
  qnr_version,
  enable,
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{qnr_id}{Questionnaire ID. GUID from server.}

\item{qnr_version}{Questionnaire version number.}

\item{enable}{Whether to enable. Values: c(TRUE, FALSE)}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{API user name}

\item{password}{API password}
}
\value{
TRUE/FALSE depending on whether operation succeeded or not.
}
\description{
Sets audio recording enabled setting for provided questionnaire. Wrapper of \code{POST /api/v1/questionnaires/{id}/{version}/recordAudio} endpoint
}
